capitulo9/geminiapi/requirements.txt
flask==3.0.3
#python-dotenv==1.0.1
google-genai
gunicorn==22.0.0          # opcional, producción
flask-cors==4.0.1         # opcional, si la vas a consumir desde un frontend